- Added Login Functionality

+++ server/controllers/users.js
@@ -28,6 +28,12 @@ const getUser = async (req, res, next) => {
   try {
     const user = await User.findById(req.params.id);
 
+    if (!userIdMatch(req)) {
+      return res.status(400).json({
+        message: `Bad Request`,
+      });
+    }
+
     if (!user) {
       res.status(404).json({
         success: true,
@@ -67,10 +73,24 @@ const createUser = async (req, res, next) => {
 const updateUser = async (req, res, next) => {
   try {
     const updatedUser = req.body;
+    if (req.body.constructor === Object && Object.keys(req.body).length === 0) {
+      return res.status(400).json({
+        message: "Bad Request",
+      });
+    }
 
-    // UserSchema.pre("save") does not work on PUT requests.
-    // You cannot access the document being updated in pre('updateOne') or pre('findOneAndUpdate') query middleware. If you need to access the document that will be updated, you need to execute an explicit query for the document.
-    // For PUT requests
+    console.log(req.user);
+
+    if (!userIdMatch(req)) {
+      return res.status(400).json({
+        message: `Bad request`,
+      });
+    }
+
+    // UserSchema.pre("save") does not work on PUT requests. Hence passwords are not getting hashed.
+    // You cannot access the document being updated in pre('updateOne') or pre('findOneAndUpdate') query middleware.
+    // If you need to access the document that will be updated, you need to execute an explicit query for the document.
+    // For PUT requests (Workaround)
     if (updatedUser.password) {
       hashedPassword = await bcrypt.hash(updatedUser.password, saltRounds);
 
@@ -81,6 +101,7 @@ const updateUser = async (req, res, next) => {
       new: true,
       runValidators: true,
     });
+
     res.status(200).json({
       success: true,
       data: user,
@@ -95,6 +116,12 @@ const updateUser = async (req, res, next) => {
 // @access  Private
 const deleteUser = async (req, res, next) => {
   try {
+    if (!userIdMatch(req)) {
+      return res.status(400).json({
+        message: "Bad Request",
+      });
+    }
+
     const user = await User.findByIdAndDelete(req.params.id);
 
     if (!user) {
@@ -110,6 +137,13 @@ const deleteUser = async (req, res, next) => {
   } catch (err) {
     console.log(err);
   }
+};
+
+// Matches User ID (req.user) with Request Parameters (params.id) to check if requesting user is loggedIn user.
+const userIdMatch = (req) => {
+  console.log(`${req.user._id}  ${req.params.id}`);
+  // req.user._id.valueOf() gives the real value of mongodb ObjectId as per MongoDB Documentation.
+  return req.user._id.valueOf().toString() === req.params.id;
 };
 
 module.exports = { getUsers, getUser, createUser, updateUser, deleteUser };
